+++ python/f2py3/UseF2Py3.cmake
@@ -74,11 +74,32 @@ macro (add_f2py3_module _name)
 
   #message(STATUS "${_name} F2PY3_Fortran_FLAGS ${F2PY3_Fortran_FLAGS}")
 
-  set(_fcompiler_opts "--fcompiler=${F2PY3_FCOMPILER}")
-  list(APPEND _fcompiler_opts "--f77exec=${CMAKE_Fortran_COMPILER}" "--f77flags='${F2PY3_Fortran_FLAGS}'")
-  if(CMAKE_Fortran_COMPILER_SUPPORTS_F90)
-     list(APPEND _fcompiler_opts "--f90exec=${CMAKE_Fortran_COMPILER}" "--f90flags='${F2PY3_Fortran_FLAGS}'")
-  endif(CMAKE_Fortran_COMPILER_SUPPORTS_F90)
+  # NOTE: This style of calling f2py is only for distutils. If you are using
+  #       Python 3.12, the backend is now meson and this will not work
+  #       so we need to test the Python version and then call the correct
+  #       f2py
+
+  if (Python3_VERSION VERSION_GREATER 3.11)
+    set(F2PY3_BACKEND "meson")
+  else ()
+    set(F2PY3_BACKEND "distutils")
+  endif ()
+
+  message(STATUS "Using F2PY3_BACKEND: ${F2PY3_BACKEND}")
+
+  if (F2PY3_BACKEND STREQUAL "distutils")
+    set(_fcompiler_opts "--fcompiler=${F2PY3_FCOMPILER}")
+    list(APPEND _fcompiler_opts "--f77exec=${CMAKE_Fortran_COMPILER}" "--f77flags='${F2PY3_Fortran_FLAGS}'")
+    if(CMAKE_Fortran_COMPILER_SUPPORTS_F90)
+      list(APPEND _fcompiler_opts "--f90exec=${CMAKE_Fortran_COMPILER}" "--f90flags='${F2PY3_Fortran_FLAGS}'")
+    endif(CMAKE_Fortran_COMPILER_SUPPORTS_F90)
+  else ()
+    set(_fcompiler_opts "")
+    list(APPEND _fcompiler_opts "--f77flags='${F2PY3_Fortran_FLAGS}'")
+    if(CMAKE_Fortran_COMPILER_SUPPORTS_F90)
+      list(APPEND _fcompiler_opts "--f90flags='${F2PY3_Fortran_FLAGS}'")
+    endif(CMAKE_Fortran_COMPILER_SUPPORTS_F90)
+  endif ()
 
   # Make the source filenames absolute.
   set(_abs_srcs)
@@ -86,6 +107,13 @@ macro (add_f2py3_module _name)
     get_filename_component(_abs_src ${_src} ABSOLUTE)
     list(APPEND _abs_srcs ${_abs_src})
   endforeach(_src ${add_f2py3_module_SOURCES})
+
+  # Let's also get all directories that the sources are in
+  set(_src_inc_dirs)
+  foreach(_src ${_abs_srcs})
+    get_filename_component(_dir ${_src} DIRECTORY)
+    list(APPEND _src_inc_dirs ${_dir})
+  endforeach(_src ${_abs_srcs})
 
   # Get a list of the include directories.
   # The f2py3 --include_paths option, used when generating a signature file,
@@ -102,6 +130,12 @@ macro (add_f2py3_module _name)
     list(APPEND _inc_dirs "${_dir}")
   endforeach(_dir)
   string(REPLACE ";" ":" _inc_paths "${_inc_dirs}")
+
+  # We also want to include the directory where the
+  # sources are located as well into _inc_opts
+  foreach(_dir ${_src_inc_dirs})
+    list(APPEND _inc_opts "-I${_dir}")
+  endforeach(_dir)
 
   set(_libs_opts)
   foreach(_lib ${add_f2py3_module_LIBRARIES})
@@ -227,6 +261,18 @@ macro (add_f2py3_module _name)
 
   # Define the command to generate the Fortran to Python3 interface module. The
   # output will be a shared library that can be imported by python.
+  # We also need to set FC in the environment to the fortran compiler
+  #message(STATUS "add_f2py3_module_SOURCES: ${add_f2py3_module_SOURCES}")
+  #message(STATUS "_inc_opts: ${_inc_opts}")
+  if ( F2PY3_BACKEND STREQUAL "meson")
+    add_custom_command(OUTPUT "${_name}${F2PY3_SUFFIX}"
+      COMMAND ${CMAKE_COMMAND} -E env "FC=${CMAKE_Fortran_COMPILER}"
+              ${F2PY3_EXECUTABLE} ${F2PY_QUIET} -m ${_name}
+              --build-dir "${CMAKE_CURRENT_BINARY_DIR}/f2py3-${_name}"
+              ${_fcompiler_opts} ${_inc_opts} -c ${_abs_srcs} ${REDIRECT_TO_DEV_NULL}
+      DEPENDS ${add_f2py3_module_SOURCES}
+      COMMENT "[F2PY3] Building Fortran to Python3 interface module ${_name}")
+  else ()
   if ( "${add_f2py3_module_SOURCES}" MATCHES "^[^;]*\\.pyf;" )
     add_custom_command(OUTPUT "${_name}${F2PY3_SUFFIX}"
       COMMAND ${F2PY3_EXECUTABLE} ${F2PY_QUIET} -m ${_name}
@@ -246,6 +292,7 @@ macro (add_f2py3_module _name)
       DEPENDS ${add_f2py3_module_SOURCES}
       COMMENT "[F2PY3] Building Fortran to Python3 interface module ${_name}")
   endif ( "${add_f2py3_module_SOURCES}" MATCHES "^[^;]*\\.pyf;" )
+  endif ()