+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updates for building with Clang on macOS
   - Add `-Wl,-ld_classic` to linker flags for all macOS
   - Add `-Wno-implicit-int` for Clang on macOS
+- Fix for using f2py and Python 3.12
 
 ### Removed